Debug heap dump

+++ src/api/linked_list.h
@@ -14,7 +14,7 @@ typedef struct node {
 } node_t;
 
 typedef struct {
-	uint32_t size;
+	uint64_t size;
 	node_t *head;
 } linked_list_t;
 
@@ -24,6 +24,6 @@ void add_node(linked_list_t *list, void *data);
 
 void remove_node(linked_list_t *list, void *data);
 
-void remove_node_at(linked_list_t *list, uint32_t index);
+void remove_node_at(linked_list_t *list, uint64_t index);
 
 #endif //TEMA1_LINKED_LIST_H
